requirements.txt
# CyberPuppy 中文網路霸凌防治系統 - 依賴包
#
# ⚠️ 重要提醒：首次安裝後請執行以下命令下載必需的大文件 (>100MB):
# python scripts/download_datasets.py
#
# 這些大文件因 GitHub 限制未包含在倉庫中，但對系統運行必不可少
# 詳細說明：docs/LARGE_FILES_SETUP.md

# Core dependencies
python-dotenv>=1.0.0
requests>=2.31.0
tqdm>=4.66.0

# Hugging Face
datasets>=2.14.0
transformers>=4.35.0
tokenizers>=0.15.0

# Data processing
pandas>=2.1.0
numpy>=1.24.0
opencc-python-reimplemented>=0.1.7  # 繁簡轉換
jieba>=0.42.1  # 中文分詞（備用）

# Machine Learning
torch>=2.1.0
scikit-learn>=1.3.0

# NLP Tools (optional but recommended)
# ckiptagger>=0.2.1  # 中研院斷詞工具（需要較大空間）

# API and Web
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
line-bot-sdk>=3.5.0  # LINE Bot SDK

# Explainability
shap>=0.44.0
captum>=0.7.0  # PyTorch 可解釋性工具

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Development tools
black>=23.12.0
flake8>=6.1.0
mypy>=1.7.0
ipykernel>=6.27.0
notebook>=7.0.0

# Optional: Perspective API client
# google-api-python-client>=2.108.0

# Logging and monitoring
loguru>=0.7.2
rich>=13.7.0  # 美化終端輸出